notifications_api/workers/sender/requirements.txt
aio-pika==8.2.1
jinja2==3.0.3
flake8==6.1.0
python-dotenv==1.0.1
pydantic==2.7.3
pydantic_core==2.18.4
pydantic-settings==2.3.0